Added graphs by SNR

+++ python_code/detectors/LSTM/lstm_detector.py
@@ -53,6 +53,6 @@ class LSTMDetector(nn.Module):
 
         if phase == 'val':
             # Decode the output
-            return torch.argmax(out, dim=2)
+            return torch.argmax(out, dim=2).float()
         else:
             return out
